Implement (de)serialize directly on ontology types
