+++ libs/@blockprotocol/type-system/crate/src/ontology/entity_type/raw.rs
@@ -1,14 +1,10 @@
-use std::{
-    collections::{HashMap, HashSet},
-    str::FromStr,
-};
+use std::{collections::HashMap, str::FromStr};
 
 use serde::{Deserialize, Serialize};
 #[cfg(target_arch = "wasm32")]
 use tsify::Tsify;
 
 use crate::{
-    ontology::entity_type::error::MergeEntityTypeError,
     raw,
     url::{BaseUrl, ParseVersionedUrlError, VersionedUrl},
     ParseEntityTypeError,
@@ -54,56 +50,6 @@ pub struct EntityType {
     pub property_object: raw::Object<raw::ValueOrArray<raw::PropertyTypeReference>>,
     #[serde(flatten)]
     pub links: raw::Links,
-}
-
-impl EntityType {
-    /// Merges another entity type into this one.
-    ///
-    /// This will:
-    ///   - remove the other entity type from the `allOf`
-    ///   - merge the `properties` and `required` fields
-    ///   - merge the `links` field
-    ///
-    /// # Notes
-    ///
-    /// - This does not validate the resulting entity type.
-    /// - The `required` field may have a different order after merging.
-    ///
-    /// # Errors
-    ///
-    /// - [`DoesNotInheritFrom`] if the other entity type is not in the `allOf` field
-    ///
-    /// [`DoesNotInheritFrom`]: MergeEntityTypeError::NotInAllOf
-    pub fn merge_parent(&mut self, other: Self) -> Result<(), MergeEntityTypeError> {
-        self.all_of.elements.remove(
-            self.all_of
-                .elements
-                .iter()
-                .position(|x| x.url == other.id)
-                .ok_or_else(|| {
-                    MergeEntityTypeError::NotInAllOf(other.id.clone(), self.id.clone())
-                })?,
-        );
-
-        self.all_of.elements.extend(other.all_of.elements);
-
-        self.property_object
-            .properties
-            .extend(other.property_object.properties);
-
-        self.property_object.required = self
-            .property_object
-            .required
-            .drain(..)
-            .chain(other.property_object.required)
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect();
-
-        self.links.links.extend(other.links.links);
-
-        Ok(())
-    }
 }
 
 impl TryFrom<EntityType> for super::EntityType {
@@ -226,23 +172,17 @@ mod tests {
             crate::test_data::entity_type::BUILDING_V1,
             None,
         );
-        let church: EntityType = check_serialization_from_str::<EntityType, raw::EntityType>(
+        let mut church: EntityType = check_serialization_from_str::<EntityType, raw::EntityType>(
             crate::test_data::entity_type::CHURCH_V1,
             None,
         );
 
-        let building_repr = raw::EntityType::from(building);
-        let mut church_repr = raw::EntityType::from(church);
-
-        church_repr
-            .merge_parent(building_repr)
+        church
+            .merge_parent(building)
             .expect("merging entity types failed");
 
-        let church_closure =
-            EntityType::try_from(church_repr).expect("entity type closure is not valid");
-
         assert!(
-            church_closure.properties().contains_key(
+            church.properties().contains_key(
                 &BaseUrl::new(
                     "https://blockprotocol.org/@alice/types/property-type/built-at/".to_owned()
                 )
@@ -250,13 +190,13 @@ mod tests {
             )
         );
         assert!(
-            church_closure.properties().contains_key(
+            church.properties().contains_key(
                 &BaseUrl::new(
                     "https://blockprotocol.org/@alice/types/property-type/number-bells/".to_owned()
                 )
                 .expect("invalid url")
             )
         );
-        assert!(church_closure.inherits_from().all_of().is_empty());
+        assert!(church.inherits_from().all_of().is_empty());
     }
 }
