+++ libs/@blockprotocol/type-system/crate/src/ontology/mod.rs
@@ -37,7 +37,8 @@ pub use shared::{
 // Re-export the repr contents so they're nicely grouped and so that they're easier to import in
 // a non-ambiguous way where they don't get confused with their non repr counterparts.
 // For example, `import crate::raw` lets you then use `raw::DataType`
-pub mod raw {
+#[allow(unreachable_pub, unused_imports)]
+pub(crate) mod raw {
     pub use super::{
         data_type::raw::{DataType, DataTypeReference},
         entity_type::{
