+++ libs/@blockprotocol/type-system/crate/src/ontology/shared/all_of/mod.rs
@@ -3,7 +3,7 @@ pub(in crate::ontology) mod raw;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct AllOf<T> {
-    elements: Vec<T>,
+    pub elements: Vec<T>,
 }
 
 impl<T> AllOf<T> {
